+++ packages/terra-breakpoints/src/terra-dev-site/doc/breakpoints/About.1.doc.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
-import { name } from '../../../../package.json';
+import { name, version } from '../../../../package.json';
 
 // Example Files
 import ActiveBreakpointProviderExample from '../example/ActiveBreakpointProviderExample';
@@ -14,6 +14,7 @@ import MixinsExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/exa
 const DocPage = () => (
   <DocTemplate
     packageName={name}
+    packageVersion={version}
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     examples={[