+++ packages/terra-list/src/terra-dev-site/doc/list/guides.7/SectionList.3.doc.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../../docs/guides/SectionList.md';
-import { name } from '../../../../../package.json';
+import { name, version } from '../../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Example Files
@@ -12,6 +12,7 @@ import SectionListSrc from '!raw-loader!../../../../../src/terra-dev-site/doc/gu
 const DocPage = () => (
   <DocTemplate
     packageName={name}
+    packageVersion={version}
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     examples={[