+++ packages/terra-show-hide/src/terra-dev-site/doc/ShowHide/ShowHide.1.doc.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
-import { name } from '../../../../package.json';
+import { name, version } from '../../../../package.json';
 
 // Component Source
 import ShowHideSrc from '!raw-loader!../../../../src/ShowHide';
@@ -24,6 +24,7 @@ import ButtonAlignRightShowHideSrc from '!raw-loader!../../../../src/terra-dev-s
 const DocPage = () => (
   <DocTemplate
     packageName={name}
+    packageVersion={version}
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     examples={[